Demo app tests: move redux helpers tests to separate files
