+++ src/shared/helpers/redux/__tests__/makeIdentityReducer.test.ts
@@ -1,0 +1,9 @@
+import { makeIdentityReducer } from '..';
+
+describe('(shared/helpers/redux) makeIdentityReducer', () => {
+  it('should return current state', () => {
+    const state = { x: 0 };
+    const identityReducer = makeIdentityReducer(state);
+    expect(identityReducer(state)).toBe(state);
+  });
+});
